archive/fastapi_aerich/requirements.txt
aerich
fastapi
fastapi_crudrouter
tortoise-orm[asyncmy]
uvicorn
python-dotenv
aiomysql
mariadb
pydantic<2.0.0
